2.1-markduplicates.sh reusability
The script now searches for input files in the STAR/ directory by extracting the basename of the *_unsort.Log.out files from 1.2-mapping.sh. The 2.2-merged_markduplicates file has been removed as it was part of a previous analysis that processed the lanes of each sample separately and then merged later on. As I already knew that the lanes could be merged, this processing merges fastq files in 0.1-lane_concat.sh